+++ CHANGELOG.md
@@ -5,6 +5,8 @@
    let property = Property.any([boolProperty1, boolProperty2, boolProperty3])
    ```
 2. Fixed Result extensions ambiguity (#733, kudos to @nekrich)
+1. Fixed Result extensions ambiguity (#733, kudos to @nekrich)
+1. Add `<~` binding operator to `Signal.Observer` (#635, kudos to @Marcocanc)
 
 # 6.0.0
 1. Dropped support for Swift 4.2 (Xcode 9)