+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # master
 *Please add new entries at the top.*
+1. add possibility to use `all` and `any` operators with array of arguments (#735, kudos to @olejnjak)
+   ```swift
+   let property = Property.any([boolProperty1, boolProperty2, boolProperty3])
+   ```
+2. Fixed Result extensions ambiguity (#733, kudos to @nekrich)
 1. Fixed Result extensions ambiguity (#733, kudos to @nekrich)
 1. Add `<~` binding operator to `Signal.Observer` (#635, kudos to @Marcocanc)